test(npm): move URL/VCS helper tests to test_utils.py
Resolves https://github.com/hermetoproject/hermeto/issues/1156

Move npm URL/VCS helper tests out of the mixed test_npm.py file
into tests/unit/package_managers/npm/test_utils.py.

Use direct imports from npm.utils so utility tests live with the
module they validate.

No behavior changes; only test relocation and import-path updates.
